requirements.txt
asgiref==3.2.10
astroid==2.4.2
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.3
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.3
django-variable-settings==0.0.1
future==0.17.1
gunicorn==19.9.0
idna==2.8
isort==5.4.2
lazy-object-proxy==1.4.3
line-bot-sdk==1.12.1
mccabe==0.6.1
psycopg2==2.8.6
pylint==2.6.0
pytz==2019.1
requests==2.22.0
six==1.15.0
sqlparse==0.3.0
static3==0.7.0
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.3
wincertstore==0.2
wrapt==1.11.2
